Installing lib reflect-metadata and rename model for models
