+++ ormconfig.json
@@ -5,6 +5,7 @@
   "username": "postgres",
   "password": "postgres",
   "database": "gostack_gobarber",
+  "entities": ["./src/models/*.ts"],
   "migrations": ["./src/database/migrations/*.ts"],
   "cli": {
     "migrationsDir": "./src/database/migrations"
