+++ package.json
@@ -13,6 +13,7 @@
     "date-fns": "^2.14.0",
     "express": "^4.17.1",
     "pg": "^8.2.1",
+    "reflect-metadata": "^0.1.13",
     "typeorm": "^0.2.25",
     "uuidv4": "^6.1.1"
   },
