+++ src/models/Appointment.ts
