+++ src/server.ts
@@ -1,3 +1,4 @@
+import 'reflect-metadata';
 import express from 'express';
 import routes from './routes';
 import './database';
